# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Estate Hub — Real Estate ERP + CRM SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List users in the company
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (company admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Deactivate a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deactivated

  # ─── LEAVE REQUESTS ──────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        clerk_user_id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - company_id
        - clerk_user_id
        - name
        - email
        - role
        - is_active
        - created_at
        - updated_at
    UserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        clerk_user_id:
          type: string
      required:
        - name
        - email
        - role
        - clerk_user_id
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        is_active:
          type: boolean

    # ─── LEAVE REQUEST SCHEMAS ───────────────────────────────
...