# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Estate Hub — Real Estate ERP + CRM SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  company_id: number;
  clerk_user_id: string;
  name: string;
  email: string;
  role: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}